Replication of McKay, Nakamura, and Steinsson (2017)

This is a replication of "The Power of Forward Guidance" by Alisdair McKay, Emi Nakamura, and Jón Steinsson.

The project was part of our evaluation for the course Numerical Methods at SciencesPo Paris in Spring 2021.

In this replication study, we do ...

HANK_MNS.params — Type

A structure that stores the parameter vector of the model.

HANK_MNS.steady_state — Type

A structure that saves the steady state of the model.

HANK_MNS.transition_CompMkts — Type

A structure that stores the complete markets transition.

HANK_MNS.transition_full — Type

A structure that stores the full model transition.

HANK_MNS.EGM — Method

EGM(c_next::Array{Float64,2},β::Float64, Rs::Array{Float64,1}, ws::Array{Float64,1},
τs::Array{Float64,1}, div::Array{Float64,1},p::params)

Conducts one iteration on consumption and labor supply using the endogenous grid method (EGM). Inputs need to be 2x1 arrays of wage, interest, etc. in current and previous period.

HANK_MNS.EGM_SS — Method

EGM_SS(c_guess::Array{Float64,2}, β::Float64,Y::Float64, p::params; T::Int=30, 
       maxit::Int = 500, tol::Float64 = 1e-7)

Computes household steady state (SS) consumption using the endogenous grid method (EGM).

HANK_MNS.aggregate_C_L — Method

aggregate_C_L(D::Array{Float64,1}, c_policies::Array{Float64,1}, R::Float64, 
              w::Float64, τ::Float64, div::Float64)

Computes aggregate consumption and labor supply of the household sector. Equivalent to expect_C in the MNS code.

HANK_MNS.check_steady_state — Method

check_steady_state(β::Float64,Y::Float64,p::params,return_distance::Bool=false)

For given β, Y and parameters, this function computes consumption and labor suppy as implied by the household decisions and returns the distance between output implied by labor and output implied by consumption as well as the distance between household asset choices and the asset target. It can be used in an Root-Finding algorithm to find the β corresponding to the asset and interest rate target. NOTE: This function is not used in the current implemenation of the model.

HANK_MNS.egm_solve_constrained — Method

egm_solve_constrained(bs::Array{Float64,1},ip::Int,w::Float64,τ::Float64,
                      div::Float64,R::Float64,p::params)

Backs out consumption level of constrained household.

HANK_MNS.find_closest_lower — Method

find_closest_lower(x::Float64,grid::Array{Float64,1})

Helper function for lineartrans. Finds the next lower asset point on the grid for savings choice x.

HANK_MNS.forward_dist — Method

forward_dist(D::Array{Float64,1},Pi::SparseMatrixCSC)

Calculates asset distribution in next period given transition matrix Pi and current distribution D.

HANK_MNS.forwardmat — Method

forwardmat(c_opt::Array{Float64,2},R::Float64,w::Float64,τ::Float64,div::Float64)

Generates a transition matrix for the aggregate wealth distribution. Uses the helper function lineartrans() as in the original code.

HANK_MNS.get_cnbp — Method

get_cnbp(xthis::Array{Float64},c::Array{Float64},R::Float64,w::Float64,
         τ::Float64,p::params,inc_idx::Int)

Helper function that, given the HH policy function and factor prices, taxes, etc., interpolates the HH policy function to the values in xthis and computes corresponding labor supply and savings.

HANK_MNS.get_figures_3_4 — Method

get_figures_3_4(;TR::Int64 = 20, T::Int64 = 200, RChange::Float64 = -0.005)

Function that replicates Figures 3 and 4 from the MNS paper. Optional argument: Time horizon of interest rate change (TR). Size of interest rate change (RChange)

HANK_MNS.get_figures_5_6 — Method

get_figures_5_6(; Horizon::StepRange{Int64,Int64} = 1:2:41, T::Int64 = 200, RChange::Float64 = -0.005)

Function to replicate Figures 5 and 6 from the MNS paper. WARNING: Running this take quite long, as a large number of transition paths are computed

HANK_MNS.get_steady_state — Method

get_steady_state(p::params,Y_guess::Float64,beta_Range::Array{Float64,1};tolβ::Float64=1e-6,tolY::Float64=1e-6)

Finds the steady state for given interest rate by iterating over β, using a simple bisection type updating algorithm. I typically used Y = 0.6 as guess for Y and betaRange = [0.95,0.99] for the baseline calibration and [0.97,0.995] for the high asset calibration, which works fine. Yguess not particularly important, I use the same as MNS.

HANK_MNS.get_table_2 — Method

get_table_2(;Horizon::Int=20,T::Int=200,RChange::Float64 = -0.005)

Function that generates the results in table 2 from the MNS paper. Returns a TypedTable.jl table containing the results. Baseline case: Interest rate change 20 periods ahead (Horizon), complete transition period length T=200, one time interest rate change (RChange) of 50 basis points

HANK_MNS.get_transition_full — Method

get_transition_full(TR::Int64,T::Int64,p::params,SS::steady_state; RChange::Float64 = -0.005)

Solves for a steady state and a transition path. Inputs: TR (time until single-quarter interest rate change), T (time horizon transition path), RChange (Change of R at time TR)

HANK_MNS.inv_dist — Method

inv_dist(Π::SparseMatrixCSC)

Find invariant steady state wealth distribution by solving for the eigenvector corresponding to the unit eigenvalue of the transition matrix. Uses insights from https://discourse.julialang.org/t/stationary-distribution-with-sparse-transition-matrix/40301

HANK_MNS.lineartrans — Method

lineartrans(bp::Array{Float64,1},par::params)

Calculates the new positions and transition probabilities on the asset grid for given savings choices. Used the to conduct a non-stochastic simulation a la Young (2010, JEDC).

HANK_MNS.logspaceshift — Method

logspaceshift(xa::Float64,xb::Float64,n::Int,x2::Float64,n_at_x2::Float64)

A function involved in making the grids. Several methods available.

HANK_MNS.makeknotd — Method

makeknotd(kmin::Float64,kmax::Float64,n::Int,logshift::Float64)

A Function involved in generating the household asset grid. Several methods available.

HANK_MNS.margU — Function

margU(c::Float64,par::params,order::Int64=1)

Helper function, computes marginal utility of consumption or second derivative of utility function (if order = 2)

HANK_MNS.reshape_c — Method

reshape_c(c::Array{Float64},p::params)

Helper function with two methods to replace par2wide and par2long from the MNS code.

HANK_MNS.set_parameters — Method

set_parameters(;β::Float64 = 0.986,    # Discount factor
                γ::Float64 = 2.0,      # Risk aversion
                ψ::Float64 = 2.0,      # Inverse Frisch elasticity
                ψ1::Float64 = 1.0,     # Labor disutility scaling (only in complete markets)
                B::Float64 = 5.5,      # Supply of Assets: 1.4 times annual GDP = 5.6 times quarterly GDP
                μ::Float64 = 1.2,      # Mark-up
                θ::Float64 = 0.15,     # Probability that Calvo fairy visits
                Rbar::Float64 = 1.005, # Target quarterly interest rate
                nz::Int64 = 3,         # Number of income grid states
                ρ::Float64 = 0.96566,            # Persistence parameter income process
                σ::Float64 = 0.01695^0.5,        # Std of income process random comp
                tax_weights::Array{Float64,1} = [0.0,0.0,1.0],   # Tax weights
                a_min::Float64  = 0.0,           # Borrowing constraint
                a_max::Float64 =75.0,            # Maximum asset holdings in grid
                x_min::Float64 = 0.001,          # Parameter to generate the consumption knots
                nk::Int64 = 1000,                # Number of grid points used for quadrature grid
                nb::Int64 = 200)

Function to construct the parameter structure.

HANK_MNS.simulate_forward — Method

simulate_forward(D0::Array{Float64,1},cpol_path::Array{Float64,2},Rpath::Array{Float64,1},
                 wpath::Array{Float64,1},div_path::Array{Float64,1},τ_path::Array{Float64,1},p::params)

Simulates a distribution of households and computes total consumption (C), labor supply (L), asset position (B) and wealth distribution (D)

HANK_MNS.simulate_step — Method

simulate_step(D::Array{Float64,1},c_pol::Array{Float64,2},R::Float64,w::Float64,
              τ::Float64,div::Float64,p::params)

Conducts forward simulation for one period. Helper function to simulateforward, equivalent to simulatestep() in MNS code.Originally used in loop in Simulateforward, replaced it to do pre-allocation there.

The following was originally in the simulate forward loop:

Cpath[t], Lpath[t], Bpath[t], Dpath[:,t+1] = simulateCLB(Dpath[:,t],reshapec(cpolpath[:,t],p) ,Rpath[t],wpath[t],τpath[t],div_path[t],p)

HANK_MNS.solve_for_transition — Method

solve_for_transition(Rpath::Array{Float64,1},wpath::Array{Float64,1},div_path::Array{Float64,1},
                     Spath::Array{Float64,1},SS::steady_state,p::params,S_tol::Float64=1e-6,w_tol::Float64=1e-6)

Solves for the perfect foresight transition path for a given interest rate path, wpath, divpath, etc. Those are guesses for the transition path. Note that guesses should include SS values as final element.

HANK_MNS.solve_for_transition_CompMkts — Method

solve_for_transition_CompMkts(Rpath, wagepath, dividendpath, Spath, stst, name, p::params)

Solves for the complete markets transition path for a given interest rate path, wpage path, dividend path, etc.

HANK_MNS.solveback — Method

solveback(c_final::Array{Float64,1},w_path::Array{Float64,1},R_path::Array{Float64,1},
          τ_path::Array{Float64},div_path::Array{Float64,1},β::Float64,p::params)

Solves backwards from a period (e.g. steady state) in which decision rules are known and given price/dividend paths. Relies on function EGM. (at the moment, this still omits β-heterogeneity). In wpath,τpath etc., the final values must correspond to the period to solve back from.

end