Replication of McKay, Nakamura, and Steinsson (2017)

This is a replication of "The Power of Forward Guidance" by Alisdair McKay, Emi Nakamura, and Jón Steinsson.

The project was part of our evaluation for the course Numerical Methods at SciencesPo Paris in Spring 2021.

In this replication study, we do ...

HANK_MNS.steady_state — Type

A structure that saves the steady state of the model.

HANK_MNS.transition_complete_markets — Type

A structure that stores the complete markets transition

HANK_MNS.transition_full — Type

A structure that stores the full model transition

HANK_MNS.EGM — Method

EGM(β::Float64,...)

Conducts one iteration on consumption and labor supply using the endogenous grid method (EGM). Inputs need to be 2x1 arrays of wage, interest etc in current and previous period

HANK_MNS.EGM_SS — Method

EGM_SS(β::Float64,Y::Float64,p::params)

Computes household steady state (SS) savings and labor supply using the endogenous grid method (EGM).

HANK_MNS.aggregate_C_L — Method

aggregateCL(D::Array{Float64,1},c_policies::Array{Float64,1},R::Float64,w::Float64,τ::Float64,div::Float64)

Computes aggregate consumption and labor supply of the household sector. Equivalent to expect_C in the MNS code.

HANK_MNS.check_steady_state — Method

check_steady_state(β::Float64,Y::Float64,p::params)

For given β,Y and Parameters, this function computes consumption and labor suppy as implied by the household decisions and returns the distance between output implied by labor and output implied by consumption as well as the distance between household asset choices and the asset target. It can be used in an Root-Finding algorithm to find the β corresponding to the asset and interest rate target.

HANK_MNS.egm_solve_constrained — Method

egm_solve_constrained()

Backs out consumption level of constrained household.

HANK_MNS.find_closest_lower — Function

findclosestlower(x::Float64,grid::Array{Float64,1})

Helper function for lineartrans

HANK_MNS.forward_dist — Method

forward_dist

Calculates Asset distribution in next period given transition matrix Pi and current distribution D

HANK_MNS.get_cnbp — Method

get_cnbp(xthis::Array{Float64},c::Array{Float64},R::Float64,
        w::Float64,τ::Float64,p::params,inc_idx::Int)

Helper function to retrieve some values

HANK_MNS.get_transition_full — Method

get_transition_full()

Solves for a steady state and a transition path. Inputs: TR (time until single-quarter interest rate change), T (time horizon transition path), RChange (Change of R at time TR)

HANK_MNS.inv_dist — Method

inv_dist

Find invariant distribution by iterating over transition matrix uses insights from https://discourse.julialang.org/t/stationary-distribution-with-sparse-transition-matrix/40301

HANK_MNS.lineartrans — Method

lineartrans(bp::Array{Float64,1},par::params)

Calculates the new positions and transition probabilities on the asset grid for given savings choices. Used the to conduct a non-stochastic simulation a la Young (2010, JEDC).

HANK_MNS.makeknotd — Method

makeknotd(kmin::Float64,kmax::Float64, n::Int, logshift::Float64)

A Function involved in generating the household asset grid. Several methods available.

HANK_MNS.margU — Function

margU(c::Float64,par::params)

Helper function, computes marginal utility of consumption or second derivative of utility function (if order = 2)

HANK_MNS.reshape_c — Method

reshape_c(c::Array{Float64},p)

helper function with two methods to replace par2wide and par2long from the MNS code

HANK_MNS.set_parameters — Method

set_parameters()

Function to construct the parameter structure.

HANK_MNS.simulate_step — Method

simulate_step()

Conducts forward simulation for one period. Helper function to simulateforward, equivalent to simulatestep() in MNS code. Originally used in loop in Simulateforward, replaced it to do pre-allocation there.

end